+++ app/components/chart/daily.js
@@ -38,6 +38,7 @@ const BolusTooltip = vizComponents.BolusTooltip;
 const SMBGTooltip = vizComponents.SMBGTooltip;
 const CBGTooltip = vizComponents.CBGTooltip;
 const FoodTooltip = vizComponents.FoodTooltip;
+const ReservoirTooltip = vizComponents.ReservoirTooltip;
 
 import Header from './header';
 import Footer from './footer';
@@ -67,6 +68,8 @@ const DailyChart = translate()(class DailyChart extends Component {
     onCBGOut: React.PropTypes.func.isRequired,
     onCarbHover: React.PropTypes.func.isRequired,
     onCarbOut: React.PropTypes.func.isRequired,
+    onReservoirHover: React.PropTypes.func.isRequired,
+    onReservoirOut: React.PropTypes.func.isRequired,
   };
 
   constructor(props) {
@@ -86,6 +89,8 @@ const DailyChart = translate()(class DailyChart extends Component {
       'onCBGOut',
       'onCarbHover',
       'onCarbOut',
+      'onReservoirHover',
+      'onReservoirOut',
     ];
 
     this.log = bows('Daily Chart');
@@ -312,6 +317,8 @@ class Daily extends Component {
                 onCBGOut={this.handleCBGOut}
                 onCarbHover={this.handleCarbHover}
                 onCarbOut={this.handleCarbOut}
+                onReservoirHover={this.handleReservoirHover}
+                onReservoirOut={this.handleReservoirOut}
                 ref="chart" />
             </div>
           </div>
@@ -376,6 +383,16 @@ class Daily extends Component {
           }}
           side={this.state.hoveredCarb.side}
           food={this.state.hoveredCarb.data}
+          bgPrefs={this.props.bgPrefs}
+          timePrefs={this.props.timePrefs}
+        />}
+        {this.state.hoveredReservoir && <ReservoirTooltip
+          position={{
+            top: this.state.hoveredReservoir.top,
+            left: this.state.hoveredReservoir.left
+          }}
+          side={this.state.hoveredReservoir.side}
+          reservoir={this.state.hoveredReservoir.data}
           bgPrefs={this.props.bgPrefs}
           timePrefs={this.props.timePrefs}
         />}
@@ -574,6 +591,29 @@ class Daily extends Component {
   handleCarbOut = () => {
     this.setState({
       hoveredCarb: false
+    });
+  };
+
+  handleReservoirHover = reservoir => {
+    var rect = reservoir.rect;
+    // range here is -12 to 12
+    var hoursOffset = sundial.dateDifference(reservoir.data.normalTime, this.state.datetimeLocation, 'h');
+    reservoir.top = rect.top + (rect.height / 2)
+    if(hoursOffset > 5) {
+      reservoir.side = 'left';
+      reservoir.left = rect.left;
+    } else {
+      reservoir.side = 'right';
+      reservoir.left = rect.left + rect.width;
+    }
+    this.setState({
+      hoveredReservoir: reservoir
+    });
+  }
+
+  handleReservoirOut = () => {
+    this.setState({
+      hoveredReservoir: false
     });
   };